activity

Records of all activities since UNHCR Programme 2016 in IATI format

Usage

activity

Format

A data frame with 825 rows and 45 variables:

iati_identifier

character A globally unique identifier for the activity.

reporting_org

character The organisation issuing the report. May be a primary source (reporting on its own activity as donor, implementing agency, etc) or a secondary source (reporting on the activities of another organisation).

reporting_org_ref

character Machine-readable identification string for the organisation issuing the report. Must be in the format RegistrationAgency-RegistrationNumber where RegistrationAgency is a valid code in the Organisation Registration Agency code list and RegistrationNumber is a valid identifier issued by the RegistrationAgency.

reporting_org_type

character The type of organisation issuing the report. See IATI codelist for values.

reporting_org_secondary_reporter

logical A flag indicating that the reporting organisation of this activity is acting as a secondary reporter. A secondary reporter is one that reproduces data on the activities of an organisation for which it is not directly responsible. This does not include a publisher officially assigned as a proxy to report on behalf of another.

title_eng

character A short, human-readable title that contains a meaningful summary of the activity. English

title_fr

character A short, human-readable title that contains a meaningful summary of the activity. French

title_es

character A short, human-readable title that contains a meaningful summary of the activity. Spanish

title_zh

character A short, human-readable title that contains a meaningful summary of the activity. Chinese

title_ru

character A short, human-readable title that contains a meaningful summary of the activity. Russian

title_ar

character A short, human-readable title that contains a meaningful summary of the activity. Arabic

description_type_1

character The type of description being provided. This is not required if only one general description of the activity is reported.

description_eng_1

character Description type 1 - English

descriptionv_fr_1

character Description type 1 - French

description_type_2

character COLUMN_DESCRIPTION

description_eng_2

character COLUMN_DESCRIPTION

descriptionv_fr_2

character COLUMN_DESCRIPTION

activity_status_code

character COLUMN_DESCRIPTION

activity_date_1

character COLUMN_DESCRIPTION

activity_date_type_1

character COLUMN_DESCRIPTION

activity_date_2

character COLUMN_DESCRIPTION

activity_date_type_2

character COLUMN_DESCRIPTION

activity_date_3

character COLUMN_DESCRIPTION

activity_date_type_3

character COLUMN_DESCRIPTION

activity_date_4

character COLUMN_DESCRIPTION

activity_date_type_4

character COLUMN_DESCRIPTION

contact_info_type

character COLUMN_DESCRIPTION

contact_info_org

character COLUMN_DESCRIPTION

contact_info_email

character COLUMN_DESCRIPTION

contact_info_website

character COLUMN_DESCRIPTION

contact_info_mailing_address

character COLUMN_DESCRIPTION

activity_scope_code

character COLUMN_DESCRIPTION

recipient_country_code

character COLUMN_DESCRIPTION

recipient_country_pct

logical COLUMN_DESCRIPTION

recipient_country

logical COLUMN_DESCRIPTION

recipient_region_code

character COLUMN_DESCRIPTION

recipient_region_vocabulary

character COLUMN_DESCRIPTION

recipient_region_vocabulary_url

character COLUMN_DESCRIPTION

recipient_region_pct

logical COLUMN_DESCRIPTION

recipient_region

character COLUMN_DESCRIPTION

collaboration_type_code

character COLUMN_DESCRIPTION

default_flow_type_code

character COLUMN_DESCRIPTION

default_finance_type_code

character COLUMN_DESCRIPTION

default_tied_status_code

character COLUMN_DESCRIPTION

capital_spend

character COLUMN_DESCRIPTION

Source

https://iatistandard.org/en/iati-standard/203/

Examples

{
head(activity,10)
}
#> # A tibble: 10 × 45
#>    iati_identifier           reporting_org     reporting_org_r… reporting_org_t…
#>    <chr>                     <chr>             <chr>            <chr>           
#>  1 XM-DAC-41121-2016         United Nations H… XM-DAC-41121     40              
#>  2 XM-DAC-41121-2016-AFR     United Nations H… XM-DAC-41121     40              
#>  3 XM-DAC-41121-2016-AME     United Nations H… XM-DAC-41121     40              
#>  4 XM-DAC-41121-2016-ASO     United Nations H… XM-DAC-41121     40              
#>  5 XM-DAC-41121-2016-EUR     United Nations H… XM-DAC-41121     40              
#>  6 XM-DAC-41121-2016-MENA    United Nations H… XM-DAC-41121     40              
#>  7 XM-DAC-41121-2016-AFR-AGO United Nations H… XM-DAC-41121     40              
#>  8 XM-DAC-41121-2016-AFR-BDI United Nations H… XM-DAC-41121     40              
#>  9 XM-DAC-41121-2016-AFR-BFA United Nations H… XM-DAC-41121     40              
#> 10 XM-DAC-41121-2016-AFR-BWA United Nations H… XM-DAC-41121     40              
#> # … with 41 more variables: reporting_org_secondary_reporter <lgl>,
#> #   title_eng <chr>, title_fr <chr>, title_es <chr>, title_zh <chr>,
#> #   title_ru <chr>, title_ar <chr>, description_type_1 <chr>,
#> #   description_eng_1 <chr>, descriptionv_fr_1 <chr>, description_type_2 <chr>,
#> #   description_eng_2 <chr>, descriptionv_fr_2 <chr>,
#> #   activity_status_code <chr>, activity_date_1 <chr>,
#> #   activity_date_type_1 <chr>, activity_date_2 <chr>, …