dataActivity

Records of all activities since UNHCR Programme 2016 in IATI format

Usage

dataActivity

Format

A data frame with 825 rows and 45 variables:

iati_identifier

character A globally unique identifier for the activity.

reporting_org

character The organisation issuing the report. May be a primary source (reporting on its own activity as donor, implementing agency, etc) or a secondary source (reporting on the activities of another organisation).

reporting_org_ref

character Machine-readable identification string for the organisation issuing the report. Must be in the format RegistrationAgency-RegistrationNumber where RegistrationAgency is a valid code in the Organisation Registration Agency code list and RegistrationNumber is a valid identifier issued by the RegistrationAgency.

reporting_org_type

character The type of organisation issuing the report. See IATI codelist for values.

reporting_org_secondary_reporter

logical A flag indicating that the reporting organisation of this activity is acting as a secondary reporter. A secondary reporter is one that reproduces data on the activities of an organisation for which it is not directly responsible. This does not include a publisher officially assigned as a proxy to report on behalf of another.

title_eng

character A short, human-readable title that contains a meaningful summary of the activity. English

title_fr

character A short, human-readable title that contains a meaningful summary of the activity. French

title_es

character A short, human-readable title that contains a meaningful summary of the activity. Spanish

title_zh

character A short, human-readable title that contains a meaningful summary of the activity. Chinese

title_ru

character A short, human-readable title that contains a meaningful summary of the activity. Russian

title_ar

character A short, human-readable title that contains a meaningful summary of the activity. Arabic

description_type_1

character The type of description being provided. This is not required if only one general description of the activity is reported.

description_eng_1

character Description type 1 - English

descriptionv_fr_1

character Description type 1 - French

description_type_2

character COLUMN_DESCRIPTION

description_eng_2

character COLUMN_DESCRIPTION

descriptionv_fr_2

character COLUMN_DESCRIPTION

activity_status_code

double COLUMN_DESCRIPTION

activity_date_1

date COLUMN_DESCRIPTION

activity_date_type_1

double COLUMN_DESCRIPTION

activity_date_2

date COLUMN_DESCRIPTION

activity_date_type_2

double COLUMN_DESCRIPTION

activity_date_3

date COLUMN_DESCRIPTION

activity_date_type_3

double COLUMN_DESCRIPTION

activity_date_4

date COLUMN_DESCRIPTION

activity_date_type_4

double COLUMN_DESCRIPTION

contact_info_type

character COLUMN_DESCRIPTION

contact_info_org

character COLUMN_DESCRIPTION

contact_info_email

character COLUMN_DESCRIPTION

contact_info_website

character COLUMN_DESCRIPTION

contact_info_mailing_address

character COLUMN_DESCRIPTION

activity_scope_code

double COLUMN_DESCRIPTION

recipient_country_code

character COLUMN_DESCRIPTION

recipient_country_pct

logical COLUMN_DESCRIPTION

recipient_country

logical COLUMN_DESCRIPTION

recipient_region_code

double COLUMN_DESCRIPTION

recipient_region_vocabulary

double COLUMN_DESCRIPTION

recipient_region_vocabulary_url

character COLUMN_DESCRIPTION

recipient_region_pct

logical COLUMN_DESCRIPTION

recipient_region

character COLUMN_DESCRIPTION

collaboration_type_code

double The type of collaboration involved in the activity’s disbursements, e.g. “bilateral” or “multilateral”.

default_flow_type_code

double COLUMN_DESCRIPTION

default_finance_type_code

double COLUMN_DESCRIPTION

default_tied_status_code

double COLUMN_DESCRIPTION

capital_spend

double COLUMN_DESCRIPTION

Source

https://iatistandard.org/en/iati-standard/203/activity-standard/iati-activities/iati-activity/

Examples

{
knitr::kable(head(dataActivity, 10))
}
#> 
#> 
#> |iati_identifier           |reporting_org                                         |reporting_org_ref | reporting_org_type|reporting_org_secondary_reporter |title_eng                                                   |title_fr                                                       |title_es |title_zh |title_ru |title_ar | description_type_1|description_eng_1                                                                                                                           |descriptionv_fr_1                                                                                                                                                | description_type_2|description_eng_2                                                                                                                                                                                   |descriptionv_fr_2                                                                                                                                                                                                                                                       | activity_status_code|activity_date_1 | activity_date_type_1|activity_date_2 | activity_date_type_2|activity_date_3 | activity_date_type_3|activity_date_4 | activity_date_type_4|contact_info_type |contact_info_org                                      |contact_info_email |contact_info_website  |contact_info_mailing_address                      | activity_scope_code|recipient_country_code |recipient_country_pct |recipient_country | recipient_region_code| recipient_region_vocabulary|recipient_region_vocabulary_url         |recipient_region_pct |recipient_region             | collaboration_type_code| default_flow_type_code| default_finance_type_code| default_tied_status_code| capital_spend|
#> |:-------------------------|:-----------------------------------------------------|:-----------------|------------------:|:--------------------------------|:-----------------------------------------------------------|:--------------------------------------------------------------|:--------|:--------|:--------|:--------|------------------:|:-------------------------------------------------------------------------------------------------------------------------------------------|:----------------------------------------------------------------------------------------------------------------------------------------------------------------|------------------:|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|--------------------:|:---------------|--------------------:|:---------------|--------------------:|:---------------|--------------------:|:---------------|--------------------:|:-----------------|:-----------------------------------------------------|:------------------|:---------------------|:-------------------------------------------------|-------------------:|:----------------------|:---------------------|:-----------------|---------------------:|---------------------------:|:---------------------------------------|:--------------------|:----------------------------|-----------------------:|----------------------:|-------------------------:|------------------------:|-------------:|
#> |XM-DAC-41121-2016         |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR - The UN Refugee Agency (2016)                        |UNHCR - L'Agence des Nations Unies pour les réfugiés (2016)    |NA       |NA       |NA       |NA       |                  1|This is the overall activity for 2016 UNHCR operations.                                                                                     |Celle-ci est l'activité globale des opérations du HCR en 2016.                                                                                                   |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-12-31      |                    4|2016-01-01      |                    2|2016-01-01      |                    1|2016-12-31      |                    3|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   1|NA                     |NA                    |NA                |                     1|                           2|NA                                      |NA                   |World                        |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-AFR     |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR operations in Africa (2016)                           |Opérations du HCR en Afrique (2016)                            |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operations in Africa for 2016 including broadly earmarked contributions.                           |Cette activité montre les détails des opérations du HCR en Afrique pour 2016, y compris les financements affectés au niveau régional.                            |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-01-01      |                    1|2016-01-01      |                    2|2016-12-31      |                    3|2016-12-31      |                    4|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   2|NA                     |NA                    |NA                |                    NA|                          99|http://www.unhcr.org/where-we-work.html |NA                   |Africa                       |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-AME     |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR operations in the Americas (2016)                     |Opérations du HCR dans les Amériques (2016)                    |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operations in the Americas for 2016 including broadly earmarked contributions.                     |Cette activité montre les détails des opérations du HCR dans les Amériques pour 2016, y compris les financements affectés au niveau régional.                    |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-01-01      |                    1|2016-01-01      |                    2|2016-12-31      |                    3|2016-12-31      |                    4|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   2|NA                     |NA                    |NA                |                    NA|                          99|http://www.unhcr.org/where-we-work.html |NA                   |The Americas                 |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-ASO     |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR operations in Asia and the Pacific (2016)             |Opérations du HCR en Asie et dans le Pacifique (2016)          |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operations in Asia and the Pacific for 2016 including broadly earmarked contributions.             |Cette activité montre les détails des opérations du HCR en Asie et dans le Pacifique pour 2016, y compris les financements affectés au niveau régional.          |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-12-31      |                    4|2016-01-01      |                    1|2016-12-31      |                    3|2016-01-01      |                    2|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   2|NA                     |NA                    |NA                |                    NA|                          99|http://www.unhcr.org/where-we-work.html |NA                   |Asia and the Pacific         |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-EUR     |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR operations in Europe (2016)                           |Opérations du HCR en Europe (2016)                             |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operations in Europe for 2016 including including broadly earmarked contributions.                 |Cette activité montre les détails des opérations du HCR en Europe pour 2016, y compris les financements affectés au niveau régional.                             |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-01-01      |                    2|2016-12-31      |                    4|2016-01-01      |                    1|2016-12-31      |                    3|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   2|NA                     |NA                    |NA                |                    NA|                          99|http://www.unhcr.org/where-we-work.html |NA                   |Europe                       |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-MENA    |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR operations in the Middle East and North Africa (2016) |Opérations du HCR au Moyen-Orient et au Afrique du Nord (2016) |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operations in the Middle East and North Africa for 2016 including broadly earmarked contributions. |Cette activité montre les détails des opérations du HCR au Moyen-Orient et au Afrique du Nord pour 2016, y compris les financements affectés au niveau régional. |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-01-01      |                    1|2016-12-31      |                    4|2016-01-01      |                    2|2016-12-31      |                    3|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   2|NA                     |NA                    |NA                |                    NA|                          99|http://www.unhcr.org/where-we-work.html |NA                   |Middle East and North Africa |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-AFR-AGO |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR Operation in Angola (2016)                            |L'opération du HCR en Angola (2016)                            |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operation in Angola for 2016 including the budget, funding, expenditure and results.               |Cette activité montre les détails de l'opération du HCR en Angola pour 2016, y compris le budget, les financements, les dépenses et les résultats.               |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-12-31      |                    4|2016-01-01      |                    2|2016-01-01      |                    1|2016-12-31      |                    3|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   4|AO                     |NA                    |NA                |                    NA|                          NA|NA                                      |NA                   |NA                           |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-AFR-BDI |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR Operation in Burundi (2016)                           |L'opération du HCR au Burundi (2016)                           |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operation in Burundi for 2016 including the budget, funding, expenditure and results.              |Cette activité montre les détails de l'opération du HCR au Burundi pour 2016, y compris le budget, les financements, les dépenses et les résultats.              |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-01-01      |                    1|2016-12-31      |                    4|2016-01-01      |                    2|2016-12-31      |                    3|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   4|BI                     |NA                    |NA                |                    NA|                          NA|NA                                      |NA                   |NA                           |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-AFR-BFA |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR Operation in Burkina Faso (2016)                      |L'opération du HCR au Burkina Faso (2016)                      |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operation in Burkina Faso for 2016 including the budget, funding, expenditure and results.         |Cette activité montre les détails de l'opération du HCR au Burkina Faso pour 2016, y compris le budget, les financements, les dépenses et les résultats.         |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-01-01      |                    1|2016-12-31      |                    4|2016-12-31      |                    3|2016-01-01      |                    2|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   4|BF                     |NA                    |NA                |                    NA|                          NA|NA                                      |NA                   |NA                           |                       4|                     10|                       110|                        5|             0|
#> |XM-DAC-41121-2016-AFR-BWA |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121      |                 40|NA                               |UNHCR Operation in Botswana (2016)                          |L'opération du HCR au Botswana (2016)                          |NA       |NA       |NA       |NA       |                  1|This activity shows details of the UNHCR operation in Botswana for 2016 including the budget, funding, expenditure and results.             |Cette activité montre les détails de l'opération du HCR au Botswana pour 2016, y compris le budget, les financements, les dépenses et les résultats.             |                  3|UNHCR, the UN Refugee Agency, is a global organisation dedicated to saving lives, protecting rights and building a better future for refugees, forcibly displaced communities and stateless people. |Le HCR, l'Agence des Nations Unies pour les réfugiés, est une organisation internationale qui a pour mission de sauver des vies, de protéger les droits des réfugiés et de construire un avenir meilleur pour les réfugiés, les communautés déplacées et les apatrides. |                    4|2016-01-01      |                    1|2016-01-01      |                    2|2016-12-31      |                    3|2016-12-31      |                    4|1                 |United Nations High Commissioner for Refugees (UNHCR) |IATI@unhcr.org     |http://www.unhcr.org/ |Case Postale 2500, CH-1211 Genève 2 Dépôt, Suisse |                   4|BW                     |NA                    |NA                |                    NA|                          NA|NA                                      |NA                   |NA                           |                       4|                     10|                       110|                        5|             0|