participating_org

An organisation involved with the activity. May be a donor, fund, agency, etc. Specifying the ref identifier is strongly recommended. May contain the organisation name as narrative. If the reporting organisation plays a role in the activity it should be repeated here. One organisation may play more than one role (eg, funding and implementing): in such a case each role should be reported and the name of the organisation repeated.

Usage

dataParticipating_org

Format

A data frame with 13990 rows and 7 variables:

iati_identifier

character A globally unique identifier for the activity.

participating_org_eng

character COLUMN_DESCRIPTION

participating_org_ref

character Machine-readable identification string for the organisation issuing the report. Must be in the format RegistrationAgency-RegistrationNumber where RegistrationAgency is a valid code in the Organisation Registration Agency code list and RegistrationNumber is a valid identifier issued by the RegistrationAgency. If this is not present then the narrative MUST contain the name of the organisation.

participating_org_type

double The type of organisation issuing the report. See IATI codelist codeOrganisationType for values.

participating_org_role

double An IATI code describing the organisation’s role in the activity (donor, agency, etc.). see codeOrganisationRole

participating_org_activity_id

logical A valid activity identifier published by the participating organisation which points to the activity that it has published to IATI that describes its role in this activity.

participating_org_crs_channel_code

logical Under CRS++ Reporting Directives this code identifies the implementing agency. Codes ending in ‘00’ are generic and are similar to the OrganisationType code.

Source

https://iatistandard.org/en/iati-standard/203/

Examples

{
knitr::kable(head(dataParticipating_org, 10))
}
#> 
#> 
#> |iati_identifier   |participating_org_eng                                 |participating_org_ref | participating_org_type| participating_org_role|participating_org_activity_id |participating_org_crs_channel_code |
#> |:-----------------|:-----------------------------------------------------|:---------------------|----------------------:|----------------------:|:-----------------------------|:----------------------------------|
#> |XM-DAC-41121-2016 |United Nations High Commissioner for Refugees (UNHCR) |XM-DAC-41121          |                     40|                      4|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |Australia for UNHCR                                   |NA                    |                     22|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |España con ACNUR                                      |NA                    |                     22|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |Japan for UNHCR                                       |NA                    |                     22|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |Sweden for UNHCR                                      |NA                    |                     22|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |UNO-Flüchtlingshilfe                                  |NA                    |                     22|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |USA for UNHCR                                         |NA                    |                     22|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |Government of Kuwait                                  |NA                    |                     10|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |Government of Portugal                                |NA                    |                     10|                      1|NA                            |NA                                 |
#> |XM-DAC-41121-2016 |Government of France                                  |NA                    |                     10|                      1|NA                            |NA                                 |