dataResult

DATASET_DESCRIPTION

Usage

dataResult

Format

A data frame

iati_identifier

character A globally unique identifier for the activity.

result_type

double COLUMN_DESCRIPTION

result_title

character COLUMN_DESCRIPTION

result_indicator_title

character COLUMN_DESCRIPTION

result_indicator_measure

double COLUMN_DESCRIPTION

result_indicator_aggregation_status

character COLUMN_DESCRIPTION

result_indicator_ascending

character COLUMN_DESCRIPTION

result_indicator_reference_code

logical COLUMN_DESCRIPTION

result_indicator_reference_uri

logical COLUMN_DESCRIPTION

result_indicator_reference_vocabulary

logical COLUMN_DESCRIPTION

result_indicator_baseline_location_ref

character COLUMN_DESCRIPTION

result_indicator_baseline_value

character COLUMN_DESCRIPTION

result_indicator_baseline_year

character COLUMN_DESCRIPTION

result_indicator_baseline_date

logical COLUMN_DESCRIPTION

result_indicator_baseline_dimension_1

character COLUMN_DESCRIPTION

result_indicator_baseline_dimension_value_1

character COLUMN_DESCRIPTION

result_indicator_baseline_dimension_2

character COLUMN_DESCRIPTION

result_indicator_baseline_dimension_value_2

character COLUMN_DESCRIPTION

result_indicator_period_start

character COLUMN_DESCRIPTION

result_indicator_period_end

character COLUMN_DESCRIPTION

result_indicator_target_value

character COLUMN_DESCRIPTION

result_indicator_target_location_ref

character COLUMN_DESCRIPTION

result_indicator_target_dimension_1

character COLUMN_DESCRIPTION

result_indicator_target_value_1

character COLUMN_DESCRIPTION

result_indicator_target_dimension_2

character COLUMN_DESCRIPTION

result_indicator_target_value_2

character COLUMN_DESCRIPTION

result_indicator_actual_value

character COLUMN_DESCRIPTION

result_indicator_actual_location_ref

character COLUMN_DESCRIPTION

result_indicator_actual_dimension_1

character COLUMN_DESCRIPTION

result_indicator_actual_value_1

character COLUMN_DESCRIPTION

result_indicator_actual_dimension_2

character COLUMN_DESCRIPTION

result_indicator_actual_value_2

character COLUMN_DESCRIPTION

result_type_name

character COLUMN_DESCRIPTION

result_type_description

character COLUMN_DESCRIPTION

indicator_measure_name

character COLUMN_DESCRIPTION

indicator_measure_description

character COLUMN_DESCRIPTION

Source

https://iatistandard.org/en/iati-standard/203/

Examples

{
knitr::kable(head(dataResult, 10))
}
#> 
#> 
#> |iati_identifier           | result_type|result_title                             |result_indicator_title                                                                                               | result_indicator_measure|result_indicator_aggregation_status |result_indicator_ascending |result_indicator_reference_code |result_indicator_reference_uri |result_indicator_reference_vocabulary |result_indicator_baseline_location_ref |result_indicator_baseline_value |result_indicator_baseline_year |result_indicator_baseline_date |result_indicator_baseline_dimension_1 |result_indicator_baseline_dimension_value_1 |result_indicator_baseline_dimension_2 |result_indicator_baseline_dimension_value_2 |result_indicator_period_start |result_indicator_period_end |result_indicator_target_value |result_indicator_target_location_ref |result_indicator_target_dimension_1 |result_indicator_target_value_1 |result_indicator_target_dimension_2 |result_indicator_target_value_2 |result_indicator_actual_value |result_indicator_actual_location_ref |result_indicator_actual_dimension_1 |result_indicator_actual_value_1 |result_indicator_actual_dimension_2 |result_indicator_actual_value_2 |result_type_name |result_type_description                                                                                                                                                     |indicator_measure_name |indicator_measure_description            |
#> |:-------------------------|-----------:|:----------------------------------------|:--------------------------------------------------------------------------------------------------------------------|------------------------:|:-----------------------------------|:--------------------------|:-------------------------------|:------------------------------|:-------------------------------------|:--------------------------------------|:-------------------------------|:------------------------------|:------------------------------|:-------------------------------------|:-------------------------------------------|:-------------------------------------|:-------------------------------------------|:-----------------------------|:---------------------------|:-----------------------------|:------------------------------------|:-----------------------------------|:-------------------------------|:-----------------------------------|:-------------------------------|:-----------------------------|:------------------------------------|:-----------------------------------|:-------------------------------|:-----------------------------------|:-------------------------------|:----------------|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:----------------------|:----------------------------------------|
#> |XM-DAC-41121-2016-AFR-AGO |           2|Civil status documentation               |% children under 12 months old who have been issued birth certificates by the authorities                            |                        2|NA                                  |NA                         |NA                              |NA                             |NA                                    |Angola                                 |30                              |2016                           |NA                             |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |35                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |NA                            |NA                                   |NA                                  |NA                              |NA                                  |NA                              |Outcome          |Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.           |Percentage             |The indicator is measured in percentages |
#> |XM-DAC-41121-2016-AFR-AGO |           2|Community mobilization                   |% of active female participants in leadership/ management structures                                                 |                        2|NA                                  |NA                         |NA                              |NA                             |NA                                    |Angola                                 |20                              |2016                           |NA                             |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |50                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |25                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |Outcome          |Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.           |Percentage             |The indicator is measured in percentages |
#> |XM-DAC-41121-2016-AFR-AGO |           1|Law and policy                           |# of advocacy interventions made                                                                                     |                        1|NA                                  |NA                         |NA                              |NA                             |NA                                    |NA                                     |NA                              |NA                             |NA                             |NA                                    |NA                                          |NA                                    |NA                                          |2016-01-01                    |2016-12-31                  |6                             |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |8                             |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |Output           |Results of the activity that came about as a direct effect of your work and specific, what is done, and what communities are reached. For example, X number of individuals. |Unit                   |The indicator is measured in units.      |
#> |XM-DAC-41121-2016-AFR-AGO |           2|Resettlement                             |% of persons of concern identified in need of resettlement submitted for resettlement                                |                        2|NA                                  |NA                         |NA                              |NA                             |NA                                    |Angola                                 |0                               |2016                           |NA                             |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |25                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |9.76                          |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |Outcome          |Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.           |Percentage             |The indicator is measured in percentages |
#> |XM-DAC-41121-2016-AFR-AGO |           2|SGBV prevention and response             |Extent community is active in SGBV prevention and survivor centered protection                                       |                        1|NA                                  |NA                         |NA                              |NA                             |NA                                    |Angola                                 |60                              |2016                           |NA                             |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |70                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |40                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |Outcome          |Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.           |Unit                   |The indicator is measured in units.      |
#> |XM-DAC-41121-2016-AFR-AGO |           2|Services for persons with specific needs |% of persons of concern with disabilities who receive services for their specific needs                              |                        2|NA                                  |NA                         |NA                              |NA                             |NA                                    |Angola                                 |46.2                            |2016                           |NA                             |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |55                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |15.32                         |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |Outcome          |Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.           |Percentage             |The indicator is measured in percentages |
#> |XM-DAC-41121-2016-AFR-AGO |           2|Solutions strategy                       |Extent Comprehensive Solutions strategy identified and agreed                                                        |                        1|NA                                  |NA                         |NA                              |NA                             |NA                                    |Angola                                 |75                              |2016                           |NA                             |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |85                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |65                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |Outcome          |Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.           |Unit                   |The indicator is measured in units.      |
#> |XM-DAC-41121-2016-AFR-AGO |           2|Status determination                     |Extent applicants have access to status determination procedure                                                      |                        1|NA                                  |NA                         |NA                              |NA                             |NA                                    |Angola                                 |20                              |2016                           |NA                             |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |35                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |4                             |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |Outcome          |Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.           |Unit                   |The indicator is measured in units.      |
#> |XM-DAC-41121-2016-AFR-BDI |           1|Child protection                         |# of best interests assessments conducted                                                                            |                        1|NA                                  |NA                         |NA                              |NA                             |NA                                    |NA                                     |NA                              |NA                             |NA                             |NA                                    |NA                                          |NA                                    |NA                                          |2016-01-01                    |2016-12-31                  |430                           |Burundi                              |Population targeted                 |Refugees and asylum-seekers     |Goal                                |Protection and mixed solutions  |471                           |Burundi                              |Population targeted                 |Refugees and asylum-seekers     |Goal                                |Protection and mixed solutions  |Output           |Results of the activity that came about as a direct effect of your work and specific, what is done, and what communities are reached. For example, X number of individuals. |Unit                   |The indicator is measured in units.      |
#> |XM-DAC-41121-2016-AFR-BDI |           1|Child protection                         |# of children's committees, groups and other structures that are operational and facilitate children's participation |                        1|NA                                  |NA                         |NA                              |NA                             |NA                                    |NA                                     |NA                              |NA                             |NA                             |NA                                    |NA                                          |NA                                    |NA                                          |2016-01-01                    |2016-12-31                  |11                            |Burundi                              |Population targeted                 |Refugees and asylum-seekers     |Goal                                |Protection and mixed solutions  |16                            |Burundi                              |Population targeted                 |Refugees and asylum-seekers     |Goal                                |Protection and mixed solutions  |Output           |Results of the activity that came about as a direct effect of your work and specific, what is done, and what communities are reached. For example, X number of individuals. |Unit                   |The indicator is measured in units.      |