transaction

Transactions recording committed or actual funds flowing in or out of an aid activity.

Usage

dataTransaction

Format

A data frame with 40419 rows and 29 variables:

iati_identifier

character COLUMN_DESCRIPTION

transaction_ref

character COLUMN_DESCRIPTION

transaction_humanitarian

logical COLUMN_DESCRIPTION

transaction_type_code

double COLUMN_DESCRIPTION

transaction_date

double COLUMN_DESCRIPTION

transaction_value_currency

character COLUMN_DESCRIPTION

transaction_value_date

double COLUMN_DESCRIPTION

transaction_value

double COLUMN_DESCRIPTION

transaction_description

character COLUMN_DESCRIPTION

transaction_provider_org_type

double COLUMN_DESCRIPTION

transaction_provider_org_ref

character COLUMN_DESCRIPTION

transaction_provider_org

character COLUMN_DESCRIPTION

transaction_aid_type_code_1

character COLUMN_DESCRIPTION

transaction_aid_type_vocabulary_1

double COLUMN_DESCRIPTION

transaction_aid_type_code_2

character COLUMN_DESCRIPTION

transaction_aid_type_vocabulary_2

double COLUMN_DESCRIPTION

transaction_value_USD

double COLUMN_DESCRIPTION

transaction_type_name

character COLUMN_DESCRIPTION

transaction_type_description

character COLUMN_DESCRIPTION

provider_org_type_name

character COLUMN_DESCRIPTION

provider_org_type_description

character COLUMN_DESCRIPTION

aid_type1_name

character COLUMN_DESCRIPTION

aid_type1_description

character COLUMN_DESCRIPTION

aid_type2_name

character COLUMN_DESCRIPTION

aid_type2_description

character COLUMN_DESCRIPTION

aid_type_vocabulary1_name

character COLUMN_DESCRIPTION

aid_type_vocabulary1_description

character COLUMN_DESCRIPTION

aid_type_vocabulary2_name

character COLUMN_DESCRIPTION

aid_type_vocabulary2_description

character COLUMN_DESCRIPTION

Source

https://iatistandard.org/en/iati-standard/203/activity-standard/iati-activities/iati-activity/transaction/

Examples

{
knitr::kable(head(dataTransaction, 10))
}
#> 
#> 
#> |iati_identifier   |transaction_ref |transaction_humanitarian | transaction_type_code|transaction_date |transaction_value_currency |transaction_value_date | transaction_value|transaction_description      | transaction_provider_org_type|transaction_provider_org_ref |transaction_provider_org |transaction_aid_type_code_1 | transaction_aid_type_vocabulary_1|transaction_aid_type_code_2 | transaction_aid_type_vocabulary_2| transaction_value_USD|transaction_type_name |transaction_type_description                                                                                                                                                  |provider_org_type_name |provider_org_type_description |aid_type1_name                                                                          |aid_type1_description                                                                                                                                                                                                                                                                                                                                                                                                 |aid_type2_name |aid_type2_description |aid_type_vocabulary1_name |aid_type_vocabulary1_description                                        |aid_type_vocabulary2_name |aid_type_vocabulary2_description                                                                            |
#> |:-----------------|:---------------|:------------------------|---------------------:|:----------------|:--------------------------|:----------------------|-----------------:|:----------------------------|-----------------------------:|:----------------------------|:------------------------|:---------------------------|---------------------------------:|:---------------------------|---------------------------------:|---------------------:|:---------------------|:-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:----------------------|:-----------------------------|:---------------------------------------------------------------------------------------|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:--------------|:---------------------|:-------------------------|:-----------------------------------------------------------------------|:-------------------------|:-----------------------------------------------------------------------------------------------------------|
#> |XM-DAC-41121-2016 |65337           |NA                       |                    11|2016-06-30       |USD                        |2016-06-30             |            995603|Monthly contribution summary |                            70|NA                           |Private donors           |B02                         |                                 1|1                           |                                 2|            995603.365|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Core contributions to multilateral institutions                                         |These funds are classified as multilateral ODA (all other categories fall under bilateral ODA). The recipient multilateral institution pools contributions so that they lose their identity and become an integral part of its financial assets. See Annex 2 of the DAC Directives for a comprehensive list of agencies core contributions to which may be reported under B02 (Section I. Multilateral institutions). |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65737           |NA                       |                    11|2016-07-31       |USD                        |2016-07-31             |             99558|Monthly contribution summary |                            70|NA                           |Private donors           |B02                         |                                 1|1                           |                                 2|             99557.520|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Core contributions to multilateral institutions                                         |These funds are classified as multilateral ODA (all other categories fall under bilateral ODA). The recipient multilateral institution pools contributions so that they lose their identity and become an integral part of its financial assets. See Annex 2 of the DAC Directives for a comprehensive list of agencies core contributions to which may be reported under B02 (Section I. Multilateral institutions). |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65228           |NA                       |                    11|2016-04-30       |USD                        |2016-04-30             |               100|Monthly contribution summary |                            70|NA                           |Private donors           |B03                         |                                 1|2                           |                                 2|                99.660|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Contributions to specific-purpose programmes and funds managed by implementing partners |In addition to their core-funded operations, international organisations, NGOs, PPPs and networks, both in provider and in third countries, set up programmes and funds with a specific sectoral, thematic or geographical focus. Donors’ bilateral contributions to such programmes and funds are recorded here.                                                                                                     |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65677           |NA                       |                    11|2016-12-31       |USD                        |2016-12-31             |              9847|Monthly contribution summary |                            70|NA                           |Private donors           |B03                         |                                 1|2                           |                                 2|              9847.020|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Contributions to specific-purpose programmes and funds managed by implementing partners |In addition to their core-funded operations, international organisations, NGOs, PPPs and networks, both in provider and in third countries, set up programmes and funds with a specific sectoral, thematic or geographical focus. Donors’ bilateral contributions to such programmes and funds are recorded here.                                                                                                     |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65476           |NA                       |                    11|2016-08-31       |USD                        |2016-08-31             |             97088|Monthly contribution summary |                            70|NA                           |Private donors           |B02                         |                                 1|1                           |                                 2|             97088.412|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Core contributions to multilateral institutions                                         |These funds are classified as multilateral ODA (all other categories fall under bilateral ODA). The recipient multilateral institution pools contributions so that they lose their identity and become an integral part of its financial assets. See Annex 2 of the DAC Directives for a comprehensive list of agencies core contributions to which may be reported under B02 (Section I. Multilateral institutions). |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65055           |NA                       |                    11|2016-01-31       |USD                        |2016-01-31             |            960637|Monthly contribution summary |                            70|NA                           |Private donors           |B02                         |                                 1|1                           |                                 2|            960637.301|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Core contributions to multilateral institutions                                         |These funds are classified as multilateral ODA (all other categories fall under bilateral ODA). The recipient multilateral institution pools contributions so that they lose their identity and become an integral part of its financial assets. See Annex 2 of the DAC Directives for a comprehensive list of agencies core contributions to which may be reported under B02 (Section I. Multilateral institutions). |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65544           |NA                       |                    11|2016-05-31       |USD                        |2016-05-31             |            953705|Monthly contribution summary |                            70|NA                           |Private donors           |B02                         |                                 1|1                           |                                 2|            953705.310|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Core contributions to multilateral institutions                                         |These funds are classified as multilateral ODA (all other categories fall under bilateral ODA). The recipient multilateral institution pools contributions so that they lose their identity and become an integral part of its financial assets. See Annex 2 of the DAC Directives for a comprehensive list of agencies core contributions to which may be reported under B02 (Section I. Multilateral institutions). |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65461           |NA                       |                    11|2016-08-31       |USD                        |2016-08-31             |              9492|Monthly contribution summary |                            70|NA                           |Private donors           |B02                         |                                 1|1                           |                                 2|              9491.715|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Core contributions to multilateral institutions                                         |These funds are classified as multilateral ODA (all other categories fall under bilateral ODA). The recipient multilateral institution pools contributions so that they lose their identity and become an integral part of its financial assets. See Annex 2 of the DAC Directives for a comprehensive list of agencies core contributions to which may be reported under B02 (Section I. Multilateral institutions). |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65476           |NA                       |                    11|2016-08-31       |USD                        |2016-08-31             |              9437|Monthly contribution summary |                            70|NA                           |Private donors           |B03                         |                                 1|2                           |                                 2|              9437.292|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Contributions to specific-purpose programmes and funds managed by implementing partners |In addition to their core-funded operations, international organisations, NGOs, PPPs and networks, both in provider and in third countries, set up programmes and funds with a specific sectoral, thematic or geographical focus. Donors’ bilateral contributions to such programmes and funds are recorded here.                                                                                                     |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |
#> |XM-DAC-41121-2016 |65387           |NA                       |                    11|2016-07-31       |USD                        |2016-07-31             |            941500|Monthly contribution summary |                            70|NA                           |Private donors           |B02                         |                                 1|1                           |                                 2|            941499.647|Incoming Commitment   |A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity. |Private Sector         |NA                            |Core contributions to multilateral institutions                                         |These funds are classified as multilateral ODA (all other categories fall under bilateral ODA). The recipient multilateral institution pools contributions so that they lose their identity and become an integral part of its financial assets. See Annex 2 of the DAC Directives for a comprehensive list of agencies core contributions to which may be reported under B02 (Section I. Multilateral institutions). |NA             |NA                    |OECD DAC                  |List of codes provided by the OECD DAC used to distinguish types of aid |Earmarking Category       |This vocabulary has been created by IATI and is derived from the Grand Bargain Earmarking Modality codelist |