dataResult

DATASET_DESCRIPTION

Usage

dataResult

Format

A data frame with 26667 rows and 24 variables:

iati_identifier

character A globally unique identifier for the activity.

result_type

character COLUMN_DESCRIPTION

result_aggregation_status

character COLUMN_DESCRIPTION

result_title

character COLUMN_DESCRIPTION

result_desc

logical COLUMN_DESCRIPTION

result_indicator_measure

character COLUMN_DESCRIPTION

result_indicator_title

character COLUMN_DESCRIPTION

result_indicator_desc

character COLUMN_DESCRIPTION

result_indicator_baseline_value

character COLUMN_DESCRIPTION

result_indicator_baseline_year

character COLUMN_DESCRIPTION

result_indicator_baseline_date

logical COLUMN_DESCRIPTION

result_indicator_baseline_location_ref

character COLUMN_DESCRIPTION

result_indicator_baseline_dimension_1

character COLUMN_DESCRIPTION

result_indicator_baseline_dimension_value_1

character COLUMN_DESCRIPTION

result_indicator_baseline_dimension_2

character COLUMN_DESCRIPTION

result_indicator_baseline_dimension_value_2

character COLUMN_DESCRIPTION

result_indicator_period_start

character COLUMN_DESCRIPTION

result_indicator_period_end

character COLUMN_DESCRIPTION

result_indicator_target_value

character COLUMN_DESCRIPTION

result_indicator_target_location_ref

character COLUMN_DESCRIPTION

result_indicator_target_dimension_1

character COLUMN_DESCRIPTION

result_indicator_target_value_1

character COLUMN_DESCRIPTION

result_indicator_target_dimension_2

character COLUMN_DESCRIPTION

result_indicator_target_value_2

character COLUMN_DESCRIPTION

Source

https://iatistandard.org/en/iati-standard/203/

Examples

{
knitr::kable(head(dataResult, 10))
}
#> 
#> 
#> |iati_identifier           |result_type |result_aggregation_status |result_title                             |result_desc |result_indicator_measure |result_indicator_title                                                                                               |result_indicator_desc                                                                   |result_indicator_baseline_value |result_indicator_baseline_year |result_indicator_baseline_date |result_indicator_baseline_location_ref |result_indicator_baseline_dimension_1 |result_indicator_baseline_dimension_value_1 |result_indicator_baseline_dimension_2 |result_indicator_baseline_dimension_value_2 |result_indicator_period_start |result_indicator_period_end |result_indicator_target_value |result_indicator_target_location_ref |result_indicator_target_dimension_1 |result_indicator_target_value_1 |result_indicator_target_dimension_2 |result_indicator_target_value_2 |
#> |:-------------------------|:-----------|:-------------------------|:----------------------------------------|:-----------|:------------------------|:--------------------------------------------------------------------------------------------------------------------|:---------------------------------------------------------------------------------------|:-------------------------------|:------------------------------|:------------------------------|:--------------------------------------|:-------------------------------------|:-------------------------------------------|:-------------------------------------|:-------------------------------------------|:-----------------------------|:---------------------------|:-----------------------------|:------------------------------------|:-----------------------------------|:-------------------------------|:-----------------------------------|:-------------------------------|
#> |XM-DAC-41121-2016-AFR-AGO |2           |0                         |Civil status documentation               |NA          |2                        |% children under 12 months old who have been issued birth certificates by the authorities                            |NA                                                                                      |30                              |2016                           |NA                             |Angola                                 |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |35                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-AGO |2           |0                         |Community mobilization                   |NA          |2                        |% of active female participants in leadership/ management structures                                                 |NA                                                                                      |20                              |2016                           |NA                             |Angola                                 |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |50                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-AGO |1           |0                         |Law and policy                           |NA          |1                        |# of advocacy interventions made                                                                                     |Related output: Advocacy conducted                                                      |NA                              |NA                             |NA                             |NA                                     |NA                                    |NA                                          |NA                                    |NA                                          |2016-01-01                    |2016-12-31                  |6                             |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-AGO |2           |0                         |Resettlement                             |NA          |2                        |% of persons of concern identified in need of resettlement submitted for resettlement                                |NA                                                                                      |0                               |2016                           |NA                             |Angola                                 |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |25                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-AGO |2           |0                         |SGBV prevention and response             |NA          |1                        |Extent community is active in SGBV prevention and survivor centered protection                                       |NA                                                                                      |60                              |2016                           |NA                             |Angola                                 |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |70                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-AGO |2           |0                         |Services for persons with specific needs |NA          |2                        |% of persons of concern with disabilities who receive services for their specific needs                              |NA                                                                                      |46.2                            |2016                           |NA                             |Angola                                 |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |55                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-AGO |2           |0                         |Solutions strategy                       |NA          |1                        |Extent Comprehensive Solutions strategy identified and agreed                                                        |NA                                                                                      |75                              |2016                           |NA                             |Angola                                 |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |85                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-AGO |2           |0                         |Status determination                     |NA          |1                        |Extent applicants have access to status determination procedure                                                      |NA                                                                                      |20                              |2016                           |NA                             |Angola                                 |Population targeted                   |Refugees and asylum seekers                 |Goal                                  |Protection and mixed solutions              |2016-01-01                    |2016-12-31                  |35                            |Angola                               |Population targeted                 |Refugees and asylum seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-BDI |1           |0                         |Child protection                         |NA          |1                        |# of best interests assessments conducted                                                                            |Related output: Best interest determination process established and operational         |NA                              |NA                             |NA                             |NA                                     |NA                                    |NA                                          |NA                                    |NA                                          |2016-01-01                    |2016-12-31                  |430                           |Burundi                              |Population targeted                 |Refugees and asylum-seekers     |Goal                                |Protection and mixed solutions  |
#> |XM-DAC-41121-2016-AFR-BDI |1           |0                         |Child protection                         |NA          |1                        |# of children's committees, groups and other structures that are operational and facilitate children's participation |Related output: Community based child protection structures established and functioning |NA                              |NA                             |NA                             |NA                                     |NA                                    |NA                                          |NA                                    |NA                                          |2016-01-01                    |2016-12-31                  |11                            |Burundi                              |Population targeted                 |Refugees and asylum-seekers     |Goal                                |Protection and mixed solutions  |